sniper-backtester-deriv/requirements.txt
pandas
matplotlib
websocket-client
